refactor(renderer): split the camera module family along its subject seams
camera-animation.ts held three subjects. Framing and first-person
navigation move out; the tween keeps the file.

- camera-framing.ts: pure pose pickers (framePointTarget,
  frameBoundsTarget, zoomExtentTarget, presetViewTarget,
  resolvePresetBounds, estimateBoundsFromPose). Pure means no back-edge
  into the animator, so no cycle -- same shape camera-fit-policy.ts
  already uses. Each framing input keeps exactly one guard, and it moves
  here with the math it guards; the animator wrappers only null-check.
- camera-first-person.ts: FirstPersonNavigator, an island that touched
  no animator field but walkVelocity. Camera owns it directly, so
  moveFirstPerson keeps its old call depth.
- camera-matrices.ts: updateCameraMatrices + computeOrthoNearFar, lifted
  off the Camera facade. Takes the shared state by reference and nothing
  else; the per-frame closure binds straight to it, so the hot path is
  the same two hops it was and allocates no more than before.
- camera-state.ts: CameraInternalState and ProjectionMode, which four
  modules needed and which lived inside one of the consumers. The import
  graph goes from a chain to a star.

camera-controls.ts is left alone: 261 code lines is already under the
house rule, and its vec3 helpers document the degradation behaviour of
their own call sites.

Fixes #2460
