+++ packages/renderer/src/camera-animation.ts
@@ -3,21 +3,36 @@
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
 
 /**
- * Camera animation system handling tweened transitions, inertia/momentum,
- * preset views, and first-person mode.
+ * Camera animation system handling tweened transitions and inertia/momentum,
+ * plus the thin wrappers that apply a pose picked by `camera-framing.ts`.
  * Extracted from Camera class using composition pattern.
+ *
+ * The tween's own failure story is the NaN *latch*: `velocity` accumulates in
+ * place and the loop spends a channel only while `Math.abs(v) > minVelocity`,
+ * which is false for NaN forever, so one non-finite gesture delta kills that
+ * channel for the session rather than for a frame (#2441/#2473). Framing's
+ * failure story is a different input class entirely and lives in
+ * `camera-framing.ts` with the guards that reject it.
  */
 
 import type { Vec3 } from './types.js';
-import { areFiniteNumbers, isUsableBounds, isUsableDistance, usableOrthoSize } from './camera-guards.js';
-import type { CameraInternalState } from './camera-controls.js';
+import { areFiniteNumbers, usableOrthoSize } from './camera-guards.js';
+import type { CameraInternalState } from './camera-state.js';
 import type { CameraControls } from './camera-controls.js';
 import type { CameraProjection } from './camera-projection.js';
+import {
+  frameBoundsTarget,
+  framePointTarget,
+  presetViewTarget,
+  resolvePresetBounds,
+  zoomExtentTarget,
+  type FramingBounds,
+} from './camera-framing.js';
 
 /**
  * Manages camera animations: tweened transitions between positions,
- * inertia/momentum after user interaction, preset view switching
- * with rotation cycling, and first-person movement.
+ * inertia/momentum after user interaction, and preset view switching
+ * with rotation cycling.
  */
 export class CameraAnimator {
   // Inertia system
@@ -37,10 +52,6 @@ export class CameraAnimator {
   private animationStartOrthoSize: number | null = null;
   private animationEndOrthoSize: number | null = null;
   private animationEasing: ((t: number) => number) | null = null;
-
-  // First-person mode
-  private isFirstPersonMode = false;
-  private walkVelocity = { x: 0, z: 0 };
 
   // Track preset view for rotation cycling (clicking same view rotates 90 degrees)
   private lastPresetView: string | null = null;
@@ -115,7 +126,7 @@ export class CameraAnimator {
 
         // Interpolate orthoSize if animating orthographic zoom.
         // The animator is the second writer that bypasses `Camera.setOrthoSize`
-        // (#2461): the read-site backstop in `updateMatrices` keeps the
+        // (#2461): the read-site backstop in `updateCameraMatrices` keeps the
         // projection matrix finite but not the state, and `getOrthoSize()`
         // reads the state — which is what a saved viewpoint persists. Keep the
         // previous half-height when the interpolation yields nothing usable.
@@ -205,32 +216,23 @@ export class CameraAnimator {
     return isAnimating;
   }
 
+  // --- Framing (pose picked in `camera-framing.ts`, applied by the tween) ---
+  //
+  // These wrappers exist on the animator because they must return the tween's
+  // promise, and the promise machinery *is* tween state. They null-check the
+  // picked target and do nothing else — in particular they do NOT re-validate
+  // the input. Each framing input is guarded exactly once, in the pure module;
+  // a second copy here would leave neither copy load-bearing, so a reverted
+  // guard would still look green.
+
   /**
    * Frame/center view on a point (keeps current distance and direction)
    * Standard CAD "Frame Selection" behavior
    */
   async framePoint(point: Vec3, duration = 300): Promise<void> {
-    // The point is added to the current offset and animated into both
-    // `position` and `target`, so a non-finite one destroys the pose. It is
-    // externally derived: `zoomToTopic` frames a BCF marker position, which is
-    // computed from a file-supplied viewpoint direction (#2461/#2466).
-    if (!areFiniteNumbers(point.x, point.y, point.z)) return;
-
-    // Keep current viewing direction and distance
-    const dir = {
-      x: this.state.camera.position.x - this.state.camera.target.x,
-      y: this.state.camera.position.y - this.state.camera.target.y,
-      z: this.state.camera.position.z - this.state.camera.target.z,
-    };
-
-    // New position: point + current offset
-    const endPos = {
-      x: point.x + dir.x,
-      y: point.y + dir.y,
-      z: point.z + dir.z,
-    };
-
-    return this.animateTo(endPos, point, duration);
+    const fit = framePointTarget(this.state, point);
+    if (!fit) return;
+    return this.animateTo(fit.position, fit.target, duration, fit.orthoSize);
   }
 
   /**
@@ -238,154 +240,44 @@ export class CameraAnimator {
    * This is what "Frame Selection" should do - zoom to fill screen
    */
   async frameBounds(min: Vec3, max: Vec3, duration = 300): Promise<void> {
-    // Bounds are the upstream input #2450 stopped short of (#2461). They are
-    // not caller-authored constants: they come from geometry, and every AABB
-    // accumulator in this package starts from `min = +Infinity, max =
-    // -Infinity` and only narrows on a comparison — which is false for a
-    // non-finite vertex — so a mesh with no finite vertices hands out that
-    // inverted sentinel as if it were a real box. `Math.max` picking the
-    // largest extent is NaN-transparent, so it reaches `position`, `target`
-    // AND `orthoSize`, and `getOrthoSize()` is what a saved viewpoint persists.
-    if (!isUsableBounds(min, max)) return;
-
-    const center = {
-      x: (min.x + max.x) / 2,
-      y: (min.y + max.y) / 2,
-      z: (min.z + max.z) / 2,
-    };
-    const size = {
-      x: max.x - min.x,
-      y: max.y - min.y,
-      z: max.z - min.z,
-    };
-    const maxSize = Math.max(size.x, size.y, size.z);
-
-    if (maxSize < 1e-6) {
-      // Very small or zero size - just center on it
-      return this.framePoint(center, duration);
-    }
-
-    // Calculate required distance based on FOV to fit bounds
-    const fovFactor = Math.tan(this.state.camera.fov / 2);
-    const distance = (maxSize / 2) / fovFactor * 1.2; // 1.2x padding for nice framing
-
-    // Get current viewing direction from view matrix (more reliable than position-target)
-    // View matrix forward is -Z axis in view space
-    const viewMatrix = this.state.viewMatrix.m;
-    // Extract forward direction from view matrix (negative Z column, normalized)
-    let dir = {
-      x: -viewMatrix[8],   // -m[2][0] (forward X)
-      y: -viewMatrix[9],   // -m[2][1] (forward Y)
-      z: -viewMatrix[10],  // -m[2][2] (forward Z)
-    };
-    const dirLen = Math.sqrt(dir.x * dir.x + dir.y * dir.y + dir.z * dir.z);
-
-    // Normalize direction
-    if (dirLen > 1e-6) {
-      dir.x /= dirLen;
-      dir.y /= dirLen;
-      dir.z /= dirLen;
-    } else {
-      // Fallback: use position-target if view matrix is invalid
-      dir = {
-        x: this.state.camera.position.x - this.state.camera.target.x,
-        y: this.state.camera.position.y - this.state.camera.target.y,
-        z: this.state.camera.position.z - this.state.camera.target.z,
-      };
-      const fallbackLen = Math.sqrt(dir.x * dir.x + dir.y * dir.y + dir.z * dir.z);
-      if (fallbackLen > 1e-6) {
-        dir.x /= fallbackLen;
-        dir.y /= fallbackLen;
-        dir.z /= fallbackLen;
-      } else {
-        // Last resort: southeast isometric
-        dir.x = 0.6;
-        dir.y = 0.5;
-        dir.z = 0.6;
-        const len = Math.sqrt(dir.x * dir.x + dir.y * dir.y + dir.z * dir.z);
-        dir.x /= len;
-        dir.y /= len;
-        dir.z /= len;
-      }
-    }
-
-    // New position: center + direction * distance
-    const endPos = {
-      x: center.x + dir.x * distance,
-      y: center.y + dir.y * distance,
-      z: center.z + dir.z * distance,
-    };
-
-    // Calculate orthoSize for orthographic mode so zoom level resets properly
-    const aspect = this.state.camera.aspect || 1;
-    const endOrthoSize = this.state.projectionMode === 'orthographic'
-      ? Math.max(0.01, maxSize / 2, maxSize / 2 / aspect) * 1.2
-      : undefined;
-
-    return this.animateTo(endPos, center, duration, endOrthoSize);
+    const fit = frameBoundsTarget(this.state, min, max);
+    if (!fit) return;
+    return this.animateTo(fit.position, fit.target, duration, fit.orthoSize);
   }
 
   async zoomExtent(min: Vec3, max: Vec3, duration = 300): Promise<void> {
-    // Same input class and same reasoning as `frameBounds` (#2461).
-    if (!isUsableBounds(min, max)) return;
-
-    const center = {
-      x: (min.x + max.x) / 2,
-      y: (min.y + max.y) / 2,
-      z: (min.z + max.z) / 2,
-    };
-    const size = {
-      x: max.x - min.x,
-      y: max.y - min.y,
-      z: max.z - min.z,
-    };
-    const maxSize = Math.max(size.x, size.y, size.z);
-
-    // Calculate required distance based on FOV
-    const fovFactor = Math.tan(this.state.camera.fov / 2);
-    const distance = (maxSize / 2) / fovFactor * 1.5; // 1.5x for padding
-
+    const fit = zoomExtentTarget(this.state, min, max);
+    if (!fit) return;
     // Update near/far planes dynamically
-    this.projection.updateNearFarPlanes(distance);
+    this.projection.updateNearFarPlanes(fit.fitDistance);
+    return this.animateTo(fit.position, fit.target, duration, fit.orthoSize);
+  }
 
-    // Keep current viewing direction
-    const dir = {
-      x: this.state.camera.position.x - this.state.camera.target.x,
-      y: this.state.camera.position.y - this.state.camera.target.y,
-      z: this.state.camera.position.z - this.state.camera.target.z,
-    };
-    const currentDistance = Math.sqrt(dir.x * dir.x + dir.y * dir.y + dir.z * dir.z);
+  /**
+   * Set preset view with explicit bounds (Y-up coordinate system)
+   * Clicking the same view again rotates 90 degrees around the view axis
+   * @param buildingRotation Optional building rotation in radians (from IfcSite placement)
+   */
+  setPresetView(
+    view: 'top' | 'bottom' | 'front' | 'back' | 'left' | 'right',
+    bounds?: FramingBounds,
+    buildingRotation?: number
+  ): void {
+    // Resolved and validated before the rotation cycles, so a rejected preset
+    // leaves the ViewCube's cycle position where it was.
+    const useBounds = resolvePresetBounds(this.state, bounds);
+    if (!useBounds) return;
 
-    // Normalize direction
-    if (currentDistance > 1e-10) {
-      dir.x /= currentDistance;
-      dir.y /= currentDistance;
-      dir.z /= currentDistance;
+    // Check if clicking the same view again - cycle rotation
+    if (this.lastPresetView === view) {
+      this.presetViewRotation = (this.presetViewRotation + 1) % 4;
     } else {
-      // Fallback direction
-      dir.x = 0.6;
-      dir.y = 0.5;
-      dir.z = 0.6;
-      const len = Math.sqrt(dir.x * dir.x + dir.y * dir.y + dir.z * dir.z);
-      dir.x /= len;
-      dir.y /= len;
-      dir.z /= len;
+      this.lastPresetView = view;
+      this.presetViewRotation = 0;
     }
 
-    // New position: center + direction * distance
-    const endPos = {
-      x: center.x + dir.x * distance,
-      y: center.y + dir.y * distance,
-      z: center.z + dir.z * distance,
-    };
-
-    // Calculate orthoSize for orthographic mode so zoom level resets properly
-    const aspect = this.state.camera.aspect || 1;
-    const endOrthoSize = this.state.projectionMode === 'orthographic'
-      ? Math.max(0.01, maxSize / 2, maxSize / 2 / aspect) * 1.5
-      : undefined;
-
-    return this.animateTo(endPos, center, duration, endOrthoSize);
+    const fit = presetViewTarget(this.state, view, useBounds, this.presetViewRotation, buildingRotation);
+    this.animateToWithUp(fit.position, fit.target, fit.up, 300);
   }
 
   /**
@@ -463,278 +355,6 @@ export class CameraAnimator {
    */
   private easeOutCubic(t: number): number {
     return 1 - Math.pow(1 - t, 3);
-  }
-
-  /**
-   * Set first-person mode
-   */
-  enableFirstPersonMode(enabled: boolean): void {
-    this.isFirstPersonMode = enabled;
-  }
-
-  /**
-   * Walk on the horizontal XZ plane (Y-up coordinate system).
-   * Forward/backward moves in the camera's horizontal facing direction.
-   * Left/right strafes perpendicular. Y position stays fixed (walking on ground).
-   * Speed scales with scene size. Movement uses smooth acceleration to avoid
-   * abrupt jumps — velocity ramps up over successive frames.
-   */
-  moveFirstPerson(forward: number, right: number, _up: number): void {
-    // Argument-side guard (#2473). `walkVelocity` is accumulated in place, so
-    // a non-finite axis latches first-person movement dead for the session
-    // exactly the way a non-finite pose did (#2441) — and it does not need an
-    // exotic value: both `moveFirstPerson(NaN, 0, 0)` and
-    // `moveFirstPerson(Infinity, 0, 0)` came back with a non-finite position
-    // on a finite pose. `_up` is unused, so it is deliberately not tested.
-    if (!areFiniteNumbers(forward, right)) return;
-
-    // Camera forward direction projected onto XZ plane
-    const dir = {
-      x: this.state.camera.target.x - this.state.camera.position.x,
-      z: this.state.camera.target.z - this.state.camera.position.z,
-    };
-    const horizLen = Math.sqrt(dir.x * dir.x + dir.z * dir.z);
-    // Finiteness, not just magnitude: `NaN < 1e-10` is false, so a malformed
-    // pose fell through here and `dir.x / horizLen` came back NaN. That does
-    // not merely produce one bad step — `walkVelocity` is accumulated in place
-    // (`+= (target - current) * 0.15`), so a single NaN frame latches it and
-    // first-person movement stays dead for the rest of the session even after
-    // the pose is corrected (#2441).
-    if (!isUsableDistance(horizLen, 1e-10)) return;
-
-    // Normalized horizontal forward and right vectors
-    const fwdX = dir.x / horizLen;
-    const fwdZ = dir.z / horizLen;
-    const rightX = -fwdZ;
-    const rightZ = fwdX;
-
-    // Target velocity from input (forward/right can be -2..2 with sprint)
-    const targetVelX = fwdX * forward + rightX * right;
-    const targetVelZ = fwdZ * forward + rightZ * right;
-
-    // Smooth acceleration: lerp current walk velocity toward target
-    this.walkVelocity.x += (targetVelX - this.walkVelocity.x) * 0.15;
-    this.walkVelocity.z += (targetVelZ - this.walkVelocity.z) * 0.15;
-
-    // Speed proportional to scene size (use camera-target distance as proxy)
-    const camDir = {
-      x: this.state.camera.position.x - this.state.camera.target.x,
-      y: this.state.camera.position.y - this.state.camera.target.y,
-      z: this.state.camera.position.z - this.state.camera.target.z,
-    };
-    const distance = Math.sqrt(camDir.x * camDir.x + camDir.y * camDir.y + camDir.z * camDir.z);
-    // The horizontal guard above does not cover this one: `camDir` includes Y,
-    // so a pose whose *only* bad coordinate is vertical reaches here with a
-    // finite `horizLen` and a NaN `distance`. `Math.max` is NaN-transparent —
-    // `Math.max(0.02, NaN)` is `NaN`, not the floor — so the clamp would
-    // forward it into the offsets written back to position and target.
-    if (!isUsableDistance(distance, 0)) return;
-    const speed = Math.max(0.02, distance * 0.004);
-
-    // Apply smoothed velocity
-    const offsetX = this.walkVelocity.x * speed;
-    const offsetZ = this.walkVelocity.z * speed;
-
-    // Move both position and target by the same offset (preserves view direction)
-    this.state.camera.position.x += offsetX;
-    this.state.camera.position.z += offsetZ;
-    this.state.camera.target.x += offsetX;
-    this.state.camera.target.z += offsetZ;
-
-    this.updateMatrices();
-  }
-
-  /**
-   * Set preset view with explicit bounds (Y-up coordinate system)
-   * Clicking the same view again rotates 90 degrees around the view axis
-   * @param buildingRotation Optional building rotation in radians (from IfcSite placement)
-   */
-  setPresetView(
-    view: 'top' | 'bottom' | 'front' | 'back' | 'left' | 'right',
-    bounds?: { min: Vec3; max: Vec3 },
-    buildingRotation?: number
-  ): void {
-    const useBounds = bounds || this.getCurrentBounds();
-    if (!useBounds) {
-      console.warn('[Camera] No bounds available for setPresetView');
-      return;
-    }
-    // Both sources need the check, not just the caller's: `getCurrentBounds()`
-    // derives its box from the live pose, so a pose that is already malformed
-    // produces a malformed box and the ViewCube would then bake it in
-    // permanently. Rejecting leaves the preset un-applied, which keeps a
-    // recoverable pose recoverable (#2461).
-    if (!isUsableBounds(useBounds.min, useBounds.max)) {
-      console.warn('[Camera] Non-finite bounds for setPresetView; keeping current view');
-      return;
-    }
-
-    // Check if clicking the same view again - cycle rotation
-    if (this.lastPresetView === view) {
-      this.presetViewRotation = (this.presetViewRotation + 1) % 4;
-    } else {
-      this.lastPresetView = view;
-      this.presetViewRotation = 0;
-    }
-
-    const center = {
-      x: (useBounds.min.x + useBounds.max.x) / 2,
-      y: (useBounds.min.y + useBounds.max.y) / 2,
-      z: (useBounds.min.z + useBounds.max.z) / 2,
-    };
-    const size = {
-      x: useBounds.max.x - useBounds.min.x,
-      y: useBounds.max.y - useBounds.min.y,
-      z: useBounds.max.z - useBounds.min.z,
-    };
-    const maxSize = Math.max(size.x, size.y, size.z);
-
-    // Calculate distance based on FOV for proper fit
-    const fovFactor = Math.tan(this.state.camera.fov / 2);
-    const distance = (maxSize / 2) / fovFactor * 1.5; // 1.5x for padding
-
-    let endPos: Vec3;
-    const endTarget = center;
-
-    // WebGL uses Y-up coordinate system internally
-    // We set both position AND up vector for proper orthogonal views
-    let upVector: Vec3 = { x: 0, y: 1, z: 0 }; // Default Y-up
-
-    // Up vector rotation options for top/bottom views (rotate around Y axis)
-    // 0: -Z, 1: -X, 2: +Z, 3: +X
-    const topUpVectors: Vec3[] = [
-      { x: 0, y: 0, z: -1 },  // 0 degrees - North up
-      { x: -1, y: 0, z: 0 },  // 90 degrees - West up
-      { x: 0, y: 0, z: 1 },   // 180 degrees - South up
-      { x: 1, y: 0, z: 0 },   // 270 degrees - East up
-    ];
-    const bottomUpVectors: Vec3[] = [
-      { x: 0, y: 0, z: 1 },   // 0 degrees - South up
-      { x: 1, y: 0, z: 0 },   // 90 degrees - East up
-      { x: 0, y: 0, z: -1 },  // 180 degrees - North up
-      { x: -1, y: 0, z: 0 },  // 270 degrees - West up
-    ];
-
-    // Apply building rotation if present (rotate around Y axis)
-    const cosR = buildingRotation !== undefined && buildingRotation !== 0 ? Math.cos(buildingRotation) : 1.0;
-    const sinR = buildingRotation !== undefined && buildingRotation !== 0 ? Math.sin(buildingRotation) : 0.0;
-
-    switch (view) {
-      case 'top': {
-        // Top view: position camera *just barely* off the +Y pole so the
-        // subsequent orbit math has a well-defined polar tangent (no pole
-        // singularity). camera.up stays world Y throughout — screen-up is
-        // then determined by lookAt projecting (0,1,0) onto perp(look),
-        // which falls along the horizontal component of `-look`.
-        //
-        // The 4 rotation cycles select which compass direction appears at
-        // the top of the screen by varying the small horizontal offset
-        // (theta in the spherical math).
-        //   rotation 0 → camera slightly to +Z of center → screen-up = +Z
-        //   rotation 1 → camera slightly to +X → screen-up = +X
-        //   rotation 2 → camera slightly to -Z → screen-up = -Z
-        //   rotation 3 → camera slightly to -X → screen-up = -X
-        // Building rotation is applied as the same Y-axis rotation that
-        // setPresetView would have used to remap the legacy up vector.
-        const poleOffset = Math.sin(0.01) * distance; // ~0.6° tilt
-        const verticalOffset = Math.cos(0.01) * distance;
-        const thetaPerRotation = [0, Math.PI / 2, Math.PI, -Math.PI / 2];
-        const thetaWorld = thetaPerRotation[this.presetViewRotation] + (buildingRotation ?? 0);
-        endPos = {
-          x: center.x + poleOffset * Math.sin(thetaWorld),
-          y: center.y + verticalOffset,
-          z: center.z + poleOffset * Math.cos(thetaWorld),
-        };
-        upVector = { x: 0, y: 1, z: 0 };
-        break;
-      }
-      case 'bottom': {
-        // Bottom view: mirror of top — phi = π − MIN_PHI.
-        const poleOffset = Math.sin(0.01) * distance;
-        const verticalOffset = Math.cos(0.01) * distance;
-        const thetaPerRotation = [Math.PI, Math.PI / 2, 0, -Math.PI / 2];
-        const thetaWorld = thetaPerRotation[this.presetViewRotation] + (buildingRotation ?? 0);
-        endPos = {
-          x: center.x + poleOffset * Math.sin(thetaWorld),
-          y: center.y - verticalOffset,
-          z: center.z + poleOffset * Math.cos(thetaWorld),
-        };
-        upVector = { x: 0, y: 1, z: 0 };
-        break;
-      }
-      case 'front':
-        // Front view: from +Z looking at model
-        // Rotate camera position around Y axis by building rotation
-        // Standard rotation: x' = x*cos - z*sin, z' = x*sin + z*cos
-        // For +Z direction (0,0,1): x' = -sin, z' = cos
-        // But we need to look at building's front, so use negative rotation
-        endPos = {
-          x: center.x + sinR * distance,
-          y: center.y,
-          z: center.z + cosR * distance,
-        };
-        upVector = { x: 0, y: 1, z: 0 }; // Y-up
-        break;
-      case 'back':
-        // Back view: from -Z looking at model
-        // For -Z direction (0,0,-1) rotated: x' = sin, z' = -cos
-        endPos = {
-          x: center.x - sinR * distance,
-          y: center.y,
-          z: center.z - cosR * distance,
-        };
-        upVector = { x: 0, y: 1, z: 0 }; // Y-up
-        break;
-      case 'left':
-        // Left view: from -X looking at model
-        // For -X direction (-1,0,0) rotated: x' = -cos, z' = sin
-        endPos = {
-          x: center.x - cosR * distance,
-          y: center.y,
-          z: center.z + sinR * distance,
-        };
-        upVector = { x: 0, y: 1, z: 0 }; // Y-up
-        break;
-      case 'right':
-        // Right view: from +X looking at model
-        // For +X direction (1,0,0) rotated: x' = cos, z' = -sin
-        endPos = {
-          x: center.x + cosR * distance,
-          y: center.y,
-          z: center.z - sinR * distance,
-        };
-        upVector = { x: 0, y: 1, z: 0 }; // Y-up
-        break;
-    }
-
-    this.animateToWithUp(endPos, endTarget, upVector, 300);
-  }
-
-  /**
-   * Get current bounds estimate (simplified - in production would use scene bounds)
-   */
-  private getCurrentBounds(): { min: Vec3; max: Vec3 } | null {
-    // Estimate bounds from camera distance
-    const dir = {
-      x: this.state.camera.position.x - this.state.camera.target.x,
-      y: this.state.camera.position.y - this.state.camera.target.y,
-      z: this.state.camera.position.z - this.state.camera.target.z,
-    };
-    const distance = Math.sqrt(dir.x * dir.x + dir.y * dir.y + dir.z * dir.z);
-    const size = distance / 2;
-
-    return {
-      min: {
-        x: this.state.camera.target.x - size,
-        y: this.state.camera.target.y - size,
-        z: this.state.camera.target.z - size,
-      },
-      max: {
-        x: this.state.camera.target.x + size,
-        y: this.state.camera.target.y + size,
-        z: this.state.camera.target.z + size,
-      },
-    };
   }
 
   /**
