+++ packages/renderer/src/camera.ts
@@ -12,9 +12,12 @@
 
 import type { Vec3, Mat4 } from './types.js';
 import { MathUtils } from './math.js';
-import { CameraControls, type CameraInternalState, type ProjectionMode } from './camera-controls.js';
+import { CameraControls } from './camera-controls.js';
+import type { CameraInternalState, ProjectionMode } from './camera-state.js';
 import { CameraAnimator } from './camera-animation.js';
 import { CameraProjection } from './camera-projection.js';
+import { FirstPersonNavigator } from './camera-first-person.js';
+import { updateCameraMatrices } from './camera-matrices.js';
 import { pickFitPolicy, type Bounds3, type FitPolicy, type PickFitPolicyOptions } from './camera-fit-policy.js';
 import {
   DEFAULT_ORTHO_SIZE,
@@ -28,6 +31,7 @@ export class Camera {
   private controls: CameraControls;
   private animator: CameraAnimator;
   private projection: CameraProjection;
+  private firstPerson: FirstPersonNavigator;
 
   constructor() {
     // Geometry is converted from IFC Z-up to WebGL Y-up during import
@@ -50,11 +54,25 @@ export class Camera {
       orbitAnchorBounds: null,
     };
 
-    const updateMatrices = () => this.updateMatrices();
+    // Bound straight to the module-level function rather than to
+    // `this.updateMatrices()`: this closure is the per-frame path every
+    // sub-system drives the matrices through, and routing it via the private
+    // method would add a hop to it. Depth is therefore unchanged from before
+    // the extraction (closure → one call), and the callee is monomorphic.
+    const updateMatrices = () => updateCameraMatrices(this.state);
     this.controls = new CameraControls(this.state, updateMatrices);
     this.projection = new CameraProjection(this.state, updateMatrices);
     this.animator = new CameraAnimator(this.state, updateMatrices, this.controls, this.projection);
+    this.firstPerson = new FirstPersonNavigator(this.state, updateMatrices);
     this.updateMatrices();
+  }
+
+  /**
+   * Cold-path convenience for this class's own setters. The hot path uses the
+   * closure built in the constructor; see the note there.
+   */
+  private updateMatrices(): void {
+    updateCameraMatrices(this.state);
   }
 
   /**
@@ -289,14 +307,14 @@ export class Camera {
    * Set first-person mode
    */
   enableFirstPersonMode(enabled: boolean): void {
-    this.animator.enableFirstPersonMode(enabled);
+    this.firstPerson.setEnabled(enabled);
   }
 
   /**
    * Move in first-person mode (Y-up coordinate system)
    */
   moveFirstPerson(forward: number, right: number, up: number): void {
-    this.animator.moveFirstPerson(forward, right, up);
+    this.firstPerson.move(forward, right, up);
   }
 
   /**
@@ -551,111 +569,5 @@ export class Camera {
    */
   getOrbitAnchorBounds(): { min: { x: number; y: number; z: number }; max: { x: number; y: number; z: number } } | null {
     return this.state.orbitAnchorBounds;
-  }
-
-  private updateMatrices(): void {
-    const dx = this.state.camera.position.x - this.state.camera.target.x;
-    const dy = this.state.camera.position.y - this.state.camera.target.y;
-    const dz = this.state.camera.position.z - this.state.camera.target.z;
-    const rawDistance = Math.sqrt(dx * dx + dy * dy + dz * dz);
-    // A non-finite position or target — a malformed BCF viewpoint reaches the
-    // public setters unvalidated — would put NaN into near/far and from there
-    // into the projection matrix, leaving the viewport dead even though
-    // `MathUtils.lookAt` scrubs the view matrix (#2441). Fall back to a
-    // neutral distance so the projection stays finite; the camera state itself
-    // is left as the caller set it.
-    const distance = Number.isFinite(rawDistance) ? rawDistance : 1;
-
-    this.state.viewMatrix = MathUtils.lookAt(
-      this.state.camera.position,
-      this.state.camera.target,
-      this.state.camera.up
-    );
-
-    if (this.state.projectionMode === 'orthographic') {
-      // Orthographic: project scene bounding sphere onto view direction for tight near/far.
-      // Tight range maximizes depth precision (less z-fighting) and prevents clipping.
-      const nf = this.computeOrthoNearFar(distance);
-      this.state.camera.near = nf.near;
-      this.state.camera.far = nf.far;
-      // Backstop for the writers that never see `setOrthoSize`'s clamp: the
-      // orthographic zoom scales `orthoSize` in place and the animator
-      // interpolates it, so a NaN reaching either (from a malformed pose, or
-      // from bounds that were themselves non-finite) would land straight in
-      // the projection matrix. Same fallback shape as the distance above.
-      const h = Number.isFinite(this.state.orthoSize) ? this.state.orthoSize : DEFAULT_ORTHO_SIZE;
-      const w = h * this.state.camera.aspect;
-      this.state.projMatrix = MathUtils.orthographicReverseZ(
-        -w, w, -h, h,
-        this.state.camera.near,
-        this.state.camera.far
-      );
-    } else {
-      // Perspective: adapt near/far based on camera-to-target distance
-      this.state.camera.near = Math.max(0.01, distance * 0.001);
-      this.state.camera.far = Math.max(distance * 10, 1000);
-      this.state.projMatrix = MathUtils.perspectiveReverseZ(
-        this.state.camera.fov,
-        this.state.camera.aspect,
-        this.state.camera.near,
-        this.state.camera.far
-      );
-    }
-
-    this.state.viewProjMatrix = MathUtils.multiply(this.state.projMatrix, this.state.viewMatrix);
-  }
-
-  /**
-   * Compute tight near/far for orthographic mode by projecting the scene
-   * bounding sphere onto the camera view direction.
-   *
-   * This gives optimal depth precision (minimizing z-fighting) while ensuring
-   * no geometry is clipped regardless of camera position or view angle.
-   */
-  private computeOrthoNearFar(distance: number): { near: number; far: number } {
-    const bounds = this.state.sceneBounds;
-    if (!bounds) {
-      // Fallback: generous range centered on camera
-      return { near: -Math.max(distance, 500), far: Math.max(distance, 500) };
-    }
-
-    // Scene bounding sphere center and radius
-    const cx = (bounds.min.x + bounds.max.x) / 2;
-    const cy = (bounds.min.y + bounds.max.y) / 2;
-    const cz = (bounds.min.z + bounds.max.z) / 2;
-    const ex = bounds.max.x - bounds.min.x;
-    const ey = bounds.max.y - bounds.min.y;
-    const ez = bounds.max.z - bounds.min.z;
-    const radius = Math.sqrt(ex * ex + ey * ey + ez * ez) / 2;
-
-    // View direction (camera looks from position toward target)
-    const pos = this.state.camera.position;
-    let vx = this.state.camera.target.x - pos.x;
-    let vy = this.state.camera.target.y - pos.y;
-    let vz = this.state.camera.target.z - pos.z;
-    const vLen = Math.sqrt(vx * vx + vy * vy + vz * vz);
-    if (vLen > 1e-8) { vx /= vLen; vy /= vLen; vz /= vLen; }
-
-    // Signed distance from camera to scene center along view direction
-    const toCenter = (cx - pos.x) * vx + (cy - pos.y) * vy + (cz - pos.z) * vz;
-
-    // Near/far as distances from camera along view direction.
-    // The sphere spans [toCenter - radius, toCenter + radius] along view dir.
-    // Add 10% padding for safety.
-    const pad = radius * 0.1 + 1;
-    let near = toCenter - radius - pad;
-    let far = toCenter + radius + pad;
-
-    // Ensure minimum range for depth precision
-    if (far - near < 1) { near -= 0.5; far += 0.5; }
-
-    // Same contract as the perspective branch: a non-finite camera position or
-    // scene bound must not reach the projection matrix (#2441).
-    if (!Number.isFinite(near) || !Number.isFinite(far)) {
-      const fallback = Math.max(Number.isFinite(distance) ? distance : 0, 500);
-      return { near: -fallback, far: fallback };
-    }
-
-    return { near, far };
   }
 }
