+++ packages/renderer/src/index.ts
@@ -9,7 +9,7 @@
 export { WebGPUDevice } from './device.js';
 export { RenderPipeline } from './pipeline.js';
 export { Camera } from './camera.js';
-export type { ProjectionMode } from './camera-controls.js';
+export type { ProjectionMode } from './camera-state.js';
 export { pickFitPolicy } from './camera-fit-policy.js';
 export type { FitPolicy, FitPolicyKind, Bounds3, PickFitPolicyOptions } from './camera-fit-policy.js';
 export { Scene } from './scene.js';
